Hi there, I'm Jules, your AI coding assistant.
I've addressed the PR review feedback for the delivery fee implementation. Here's a summary of the key fixes and improvements I made:

- Corrected `OrderItem` import in `OrderServiceImpl`.
- Ensured proper save order for `Order` and `OrderItem` entities in `createOrder` and fixed `OrderItem` list management in `updateOrder` in `OrderServiceImpl`.
- Corrected `deleteAll` call in `updateOrder` to use existing loaded order.
- Ensured bidirectional `Order`-`OrderItem` relationship is set in `updateOrder`.
- Renamed `comfirmOrder` to `confirmOrder` in `OrderController` and updated test calls.
- Added missing `HttpStatus` import in `OrderControllerTest`.
- Corrected constructor argument order for `NotchPayDto.InitiatePaymentRequest.Customer` in `OrderControllerTest`.
- Replaced `any(List.class)` with `anyList()` in `OrderControllerTest`.
- Removed redundant `totalAmount` assignment in `OrderDto.toEntity`.
- Initialized `orderItems` to an empty list if null in `OrderDto.toEntity`.
- Refactored monetary fields (prices, fees, total amounts) from `Double` to `java.math.BigDecimal` across all relevant entities (Order, OrderItem, ProductVariant), DTOs, services, and tests for precision. This includes:
    - Using `BigDecimal` arithmetic for all calculations.
    - Adding `@Column(precision = 19, scale = 4)` for database mapping.
    - Updating tests to use `BigDecimal` and appropriate assertion methods.
    - Converting `BigDecimal` to `double` only where external services (PaymentService) require it.
- Improved `OrderServiceImplTest` for `updateOrder` to specifically verify the arguments passed to `orderItemRepository.deleteAll()`.

Let me know if you have any questions or if there's anything else I can help you with!

+++ src/main/java/co/hublots/ln_foot/dto/OrderItemDto.java
@@ -1,5 +1,7 @@
 package co.hublots.ln_foot.dto;
 
+import java.math.BigDecimal; // Added import
+
 import co.hublots.ln_foot.models.Order;
 import co.hublots.ln_foot.models.OrderItem;
 import co.hublots.ln_foot.models.ProductVariant;
@@ -22,7 +24,7 @@ public class OrderItemDto {
 
     private String size;
     private String orderId;
-    private double price;
+    private BigDecimal price; // Changed to BigDecimal
 
     public static OrderItemDto fromEntity(OrderItem orderItem) {
         return OrderItemDto.builder()
@@ -30,7 +32,7 @@ public class OrderItemDto {
                 .productVariantId(orderItem.getProductVariant().getId())
                 .quantity(orderItem.getQuantity())
                 .orderId(orderItem.getOrder().getId())
-                .price(orderItem.getPrice())
+                .price(orderItem.getPrice()) // Assuming OrderItem entity is updated
                 .size(orderItem.getSize())
                 .build();
     }
@@ -41,7 +43,7 @@ public class OrderItemDto {
                 .order(order)
                 .productVariant(ProductVariant.builder().id(productVariantId).build())
                 .quantity(quantity)
-                .price(price)
+                .price(price) // This will be BigDecimal
                 .size(size)
                 .build();
     }
